refactor(farm): fix shareURL being undefined by moving ShareModal to ProMMFarm
